[mlir] Fix two build warnings (NFC)
/data/llvm-project/mlir/lib/Dialect/Tensor/Utils/Utils.cpp:62:11: error: comparison of integers of different signs: 'int64_t' (aka 'long') and 'size_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
  if (dim >= shape.size())
      ~~~ ^  ~~~~~~~~~~~~
1 error generated.

/data/llvm-project/mlir/lib/Dialect/Tensor/IR/TensorTilingInterfaceImpl.cpp:484:8: error: unused variable 'appendIndex' [-Werror,-Wunused-variable]
  auto appendIndex = [&](Value val, SmallVector<Value> &dynIndices,
       ^
1 error generated.

+++ mlir/lib/Dialect/Tensor/Utils/Utils.cpp
@@ -59,7 +59,7 @@ FailureOr<OpFoldResult> mlir::tensor::createDimValue(OpBuilder &b, Location loc,
   if (!tensorTy)
     return failure();
   auto shape = tensorTy.getShape();
-  if (dim >= shape.size())
+  if (dim >= static_cast<int64_t>(shape.size()))
     return failure();
   if (ShapedType::isDynamic(shape[dim]))
     return OpFoldResult(b.createOrFold<tensor::DimOp>(loc, rankedTensor, dim));
